Refactoring mapJestToTestAdapter into a class, using full path as test ID, changing test ID to be more readable for logs
